android_tools_version_fix: android tools version bump
using android version 3.4.0 to avoid compilation errors when using proguard.
Android tools  below version 3.4.0 will throw a compilation error related with warnings due to classes not found.
In the case of our SDK, those classes belong the TPN SDK and it's expected for those classes to not all be present at compile time.

3.4.0 is not the latest android tools version but is guaranteed to work across Supported Unity version within the current gradle 'mainTemplate'